refactor: add mesh `shift` function
When physical ball traveling, the position of ball are changed, we use `shift`
function to update it's position. In the meanwhile, the mesh of that ball also
need to be shifted to correctly show ball in the canvas when the world try to
draw it.

Before we let world do the mesh shift operation, now we capsule package it as
`shift` function similar to all the shapes.

Signed-off-by: iceserval <icelingshan@foxmail.com>